Add utility functions with tests usting jest
Adds two utility function, with tests using jest, to create a string representation of a
number in imperial format with feet, inches, and fractional part. Uses this utility function
inside Component classes when displaying lengths.
Adds utility function, with tests using jest, to create a string representation of a number
with commas, given the number.
Adds utility function to create a string representation of a time duration with days, hours,
minutes, seconds, and milliseconds.
Adds utility function to create a number given separate number of feet, inches, and
fraction of inches.